Updating java docs and formatting changes
